@@ -3,10 +3,9 @@ import { Component, OnInit, ChangeDetectionStrategy, ChangeDetectorRef, ViewChil
 import { OverlayContainer } from '@angular/cdk/overlay';
 
 import { GUIGlobal } from '../../providers/GUIGlobal';
-import { MatGridList, MatGridTile } from '@angular/material';
-import { NbTabsetComponent } from '@nebular/theme/components/tabset/tabset.component';
+import { MatGridList, MatGridTile } from '@angular/material/grid-list';
 import { NbSelectComponent } from '@nebular/theme/components/select/select.component';
-import { NbDialogService } from '@nebular/theme';
+import {NbDialogService, NbTabsetComponent} from '@nebular/theme';
 import { ColorPickerModule } from 'ngx-color-picker';
 import { ngfModule, ngf } from "angular-file";
 
@@ -25,10 +24,12 @@ import { TextInputWindow } from './textInputWindow/textInputWindow.component';
 })
 export class GeneratorComponent implements OnInit {
 
+  selectedItem = '2';
+
   tooltipComponent = GUITooltip;
 
-  @ViewChild('refTabSet') tabSet: NbTabsetComponent;
-  @ViewChild('refTabFooter') tabSetFooter: NbTabsetComponent;
+  @ViewChild('refTabSet', { static: false }) tabSet: NbTabsetComponent;
+  @ViewChild('refTabFooter', { static: false }) tabSetFooter: NbTabsetComponent;
   activeTab: string = "";
   activeFooterTab: string = "";
   settingsLocked: boolean = false;
@@ -547,7 +548,7 @@ export class GeneratorComponent implements OnInit {
               this.cd.markForCheck();
               this.cd.detectChanges();
 
-              refPresetSelect.setSelected = trimmedName;
+              refPresetSelect.selected = trimmedName;
 
               //console.log("Preset created");
             }
