@@ -3,7 +3,7 @@ import { NbMenuItem } from '@nebular/theme';
 export const MENU_ITEMS: NbMenuItem[] = [
   {
     title: 'Generator',
-    icon: 'nb-home',
+    icon: 'nb-home', // TODO Check maybe remove nb
     link: '/pages/generator',
     home: true,
   }
